refactor: 클래스 및 파일구조 정리 #24
