+++ mobile/Baetaverse/Baetaverse/Network/APIModel/RequestModel/SignUpRequest.swift
