+++ mobile/Baetaverse/Baetaverse/Network/APIModel/ResponseModel/LoginResponse.swift
