+++ mobile/Baetaverse/Baetaverse/Network/APIModel/ResponseModel/SignUpResponse.swift
