+++ mobile/Baetaverse/Baetaverse/Network/Implements/APIRequestImplement.swift
@@ -1,34 +1,11 @@
 //
-//  APIRequest.swift
+//  APIRequestImplement.swift
 //  Baetaverse
 //
 //  Created by JeongTaek Han on 2022/03/29.
 //
 
 import Foundation
-
-protocol APIRequest {
-    
-    var httpMethod: HTTPMethod { get }
-    var baseURL: String { get }
-    var path: String { get }
-    var query: [String: String] { get }
-    var headers: [String: String] { get }
-    
-}
-
-protocol APIRequestGET: APIRequest {
-    
-    var parameters: [String: String] { get }
-    
-}
-
-protocol APIRequestPOST: APIRequest {
-    
-    var contentType: String { get }
-    var body: Data? { get }
-    
-}
 
 extension APIRequest {
