+++ mobile/Baetaverse/Baetaverse/Network/Interface/APIRequestHostInfo.swift
