+++ mobile/Baetaverse/Baetaverse/Service/APIServiceConfig/BaetaverseAPIRequestHostInfo.swift
