--- mobile/Baetaverse/Baetaverse/ViewController/ViewController.swift
@@ -1,19 +1,0 @@
-//
-//  ViewController.swift
-//  Baetaverse
-//
-//  Created by JeongTaek Han on 2022/03/29.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-    }
-
-
-}
-
